+++ PSNewScriptFile/internal/functions/Initialize-NewScript.ps1
@@ -1,13 +1,10 @@
 ﻿function Initialize-NewScript {
     <#
     .SYNOPSIS
-        Helper function for generating a script files
+        Function for generating a script files
 
     .DESCRIPTION
         This helper function will generate the script context and build the .ps1 file
-
-    .EXAMPLE
-        C:\> New-ScriptFile
 
     .NOTES
         This is where all the magic happens
@@ -18,6 +15,7 @@
         [string]
         $ScriptName,
 
+        [parameter(HelpMessage = "Override value of users save location, otherwise use the stored configuration setting.")]
         [string]
         $FilePath,
 
@@ -28,10 +26,70 @@
         $Description,
 
         [string]
+        $Example,
+
+        [string]
         $Notes,
 
         [Int32]
         $NumberOfParameters = 1,
+
+        [switch]
+        $Alias,
+
+        [switch]
+        $AllowNull,
+
+        [switch]
+        $AllowEmptyString,
+
+        [switch]
+        $AllowEmptyCollection,
+
+        [switch]
+        $HelpMessage,
+
+        [switch]
+        $Mandatory,
+
+        [switch]
+        $Position,
+
+        [switch]
+        $ParameterSet,
+
+        [switch]
+        $ValueFromPipeline,
+
+        [switch]
+        $ValueFromPipelineByPropertyName,
+
+        [switch]
+        $ValueFromRemainingArguements,
+
+        [switch]
+        $ValidateNotNull,
+
+        [switch]
+        $ValidateNotNullOrEmpty,
+
+        [switch]
+        $ValidateCount,
+
+        [switch]
+        $ValidateLength,
+
+        [switch]
+        $ValidatePattern,
+
+        [switch]
+        $ValidateRange,
+
+        [switch]
+        $ValidateSet,
+
+        [switch]
+        $ValidateScript,
 
         [bool]
         $EnableException
@@ -39,6 +97,17 @@
 
     begin {
         $index = 0
+        $validationParameters = $PSBoundParameters
+        $propertiesToRemove = @(
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.ScriptName"),
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.FilePath"),
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.Synopsis"),
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.Description"),
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.Notes"),
+            (Get-PSFConfigValue -FullName "PSNewScriptFile.Configuration.Property.NumberOfParameters")
+        )
+
+        foreach ($removeProp in $propertiesToRemove) { [void]$validationParameters.Remove("$removeProp") }
     }
     process {
         [System.Collections.ArrayList]$finalScriptFile = @()
@@ -64,20 +133,20 @@
 
             $finalScriptFile += ("`n`t.EXAMPLE")
             if (-NOT $Example) {
-                if($NumberOfParameters -eq 1){ $finalScriptFile += "`t`tC:\>$ScriptName.ps1 -Parameter Arguement" }
-                elseif($NumberOfParameters -gt 1) {
-                $index = 0
-                while ($index -lt $NumberOfParameters) {
-                    $index ++
-                    $tempString += "-Parameter Arguement "
+                if ($NumberOfParameters -eq 1) { $finalScriptFile += "`t`tC:\> $ScriptName.ps1 -Parameter Arguement" }
+                elseif ($NumberOfParameters -gt 1) {
+                    $index = 0
+                    while ($index -lt $NumberOfParameters) {
+                        $index ++
+                        $tempString += "-Parameter Arguement "
+                    }
+                    $finalScriptFile += "`t`tC:\>$ScriptName.ps1 $tempString"
                 }
-                $finalScriptFile += "`t`tC:\>$ScriptName.ps1 $tempString"
-            }
             }
             else { $finalScriptFile += "`t`t$Example" }
 
             $finalScriptFile += "`n`t.NOTES"
-            if (-NOT $Notes) { $finalScriptFile += "`t`tPlease add your notes here!" }
+            if (-NOT $Notes) { $finalScriptFile += "`t`tPlease add your notes here." }
             else { $finalScriptFile += "`t`t$Notes" }
             $finalScriptFile += "`t#>"
             #endregion_Notes
@@ -85,9 +154,13 @@
             #region_paramaters
             $finalScriptFile += "`n`t[CmdletBinding()]`n`tparam("
 
+            #region_Construct_Parameters_Output
             $index = 0  # Reset counter
             if ($NumberOfParameters -eq 1) {
-                $finalScriptFile += "`t[Object]"
+                if ($validationParameters.Count -gt 0) {
+                    foreach ($item in $validationParameters.Keys) { $finalScriptFile += Add-ParametersToScriptfile -Parameter $item }
+                }
+                $finalScriptFile += "`t[ObjectType]"
                 $finalScriptFile += "`t`$parameter$index"
                 $finalScriptFile += "`t)"
             }
@@ -95,15 +168,18 @@
                 $index = 0
                 while ($index -lt $NumberOfParameters) {
                     $index ++
-                    $finalScriptFile += "`t[Object]"
-                    if ($index -ne $NumberOfParameters) {
-                        $finalScriptFile += "`t`$parameter$index,"
-                    }
-                    else { $finalScriptFile += "`t`$parameter$index" }
-                }
-                $finalScriptFile += "`t)"
-            }
+                    if ($validationParameters.Count -gt 0) {
 
+                        $finalScriptFile += Add-ParametersToScriptfile -ParameterList $validationParameters
+                        $finalScriptFile += "`t[ObjectType]"
+                        if ($index -ne $NumberOfParameters) { $finalScriptFile += "`t`$parameter$index,`n" }
+                        else { $finalScriptFile += "`t`$parameter$index`n" }
+                    }
+                }
+            }
+            #endregion_Construct_Parameters_Output
+
+            $finalScriptFile += "`t)"
             $finalScriptFile += "`n`tbegin{}`n`tprocess{}`n`tend{}"
             $finalScriptFile += "`n}"
             #endregion_paramaters