+++ PSNewScriptFile/changelog.md
@@ -1,10 +1,26 @@
 ﻿# Changelog
+
+## 1.0.1 (2020-07-27)
+
+- Upd: Changes to New-PSScript.ps1
+- Upd: Reordered parameters
+- Upd: Added all validation parameters that PowerShell uses
+- Upd: Fixed formatting violations in Changelog.md
+- New: Added new script parameters
+- New: configuration.properties.ps1 file
+- New: Add-ParametersToScriptfile.ps1 file
+- New: Script for building scripted template with all properties requested
+- Upd: Changes to Initialize-NewScript.ps1
+- New: Remoevd example (this is an internal function)
+- New: Added new code to check for properties to be added and removed for both parameter and attribute lists
+
 ## 1.0.0 (2020-07-26)
- - New: New-PSScript.ps1
- - New: Initialize-NewScript.ps1
- - New: Save-PSScript.ps1
- - Upd: Configuration.ps1
- - Upd: Changlog.md
- - Upd: README.md
- - Upd: strings.psd1
- - Upd: PSNewScriptFile.psd1
+
+- New: New-PSScript.ps1
+- New: Initialize-NewScript.ps1
+- New: Save-PSScript.ps1
+- Upd: Configuration.ps1
+- Upd: Changlog.md
+- Upd: README.md
+- Upd: strings.psd1
+- Upd: PSNewScriptFile.psd1